# backend/05_APIs_Externas/api_rest/openapi.yaml
openapi: 3.0.3
info:
  title: METGO 3D API REST
  description: API agrometeorológica MVP — Valle de Aconcagua (Chile)
  version: 1.0.0-mvp
servers:
  - url: http://127.0.0.1:8080
    description: Desarrollo local
  - url: https://metgo-api.onrender.com
    description: Producción Render
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login JWT (E9 — claim sitio)
      description: |
        Emite JWT con `sub`, `role`, `tenant` (legacy) y **`sitio`** (producto).
        El body puede incluir `sitio` para scoped login; un usuario de Quillota
        no puede pedir `mantos_blancos` (403). Admin global (`sitio=null` en membresía)
        puede omitir sitio (acceso a todos) o pedirlo para limitar el token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: admin }
                password: { type: string, example: admin123 }
                sitio: { type: string, example: quillota, description: "Producto METGO (opcional)" }
      responses:
        "200":
          description: Token emitido (user.sitio en payload)
        "401":
          description: Credenciales incorrectas
        "403":
          description: Sitio no autorizado para el usuario
  /api/auth/register:
    post:
      tags: [auth]
      summary: Registro de usuario (rol lectura, auto-registro demo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: campo_norte }
                password: { type: string, minLength: 6 }
                email: { type: string, format: email }
                sitio: { type: string, example: quillota }
      responses:
        "201":
          description: Usuario creado y token emitido
        "400":
          description: Validación fallida
  /api/auth/me:
    get:
      tags: [auth]
      summary: Usuario actual (JWT)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: "{ username, role, tenant, sitio }"
        "401":
          description: Sin token
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Renovar JWT preservando sitio
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Nuevo access_token
        "401":
          description: Sin token
  /api/me/preferencias:
    get:
      tags: [auth]
      summary: Preferencias/favoritos del usuario para un sitio
      security: [{ bearerAuth: [] }]
      parameters:
        - name: sitio
          in: query
          schema: { type: string, default: quillota }
      responses:
        "200":
          description: "{ usuario, sitio, prefs, favorites }"
        "403":
          description: Sitio no autorizado
    put:
      tags: [auth]
      summary: Guardar preferencias/favoritos (usuario+sitio)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sitio: { type: string }
                prefs: { type: object }
                favorites: { type: array }
      responses:
        "200":
          description: Preferencias guardadas
        "403":
          description: Sitio no autorizado
        "503":
          description: Supabase no disponible
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT